microservices/questionnaire_llm_updater/requirements.txt
paho-mqtt==1.6.1
SQLAlchemy==2.0.18
psycopg2-binary==2.9.5
pyspark==3.5.0
requests==2.28.2
requests-oauthlib==1.3.1